+++ AlpacaService/Release-Notes.md
@@ -10,6 +10,11 @@ BN.PROJECT is a platform which combines brokerage operations with algorithmic tr
 - implement code coverage tool coverlet
 - implement code coverage report generator 
 
+## Version 2024.10.2
+
+- test result storage in db
+- implement api for backtest results
+
 ## Version 2024.10.1
 
 - implement multi user support for backtests